requirements_ifr.txt
accelerate==1.1.1
datasets==2.14.6
decorator==5.1.1
decord==0.6.0
gspread==6.1.4
huggingface==0.0.1
huggingface-hub==0.26.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
numpy
nvitop==1.4.2
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
openai==0.27.0
opencv-python==4.7.0.72
pandas
peft==0.5.0
Pillow==10.0.1
pip==23.3
pycocoevalcap==1.2
pycocotools==2.0.8
regex==2022.10.31
requests==2.31.0
requests-oauthlib==2.0.0
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==3.4.1
setuptools==68.0.0
six==1.16.0
spacy==3.5.1
timm==1.0.13
tokenizers==0.21.0
tqdm==4.64.1
traitlets==5.11.2
transformers==4.49.0
triton==2.1.0
typer==0.7.0
